[PIL-1528-update-api] Update API to include error response schemas
[PIL-1528-update-api] Working documentation
