+++ app/uk/gov/hmrc/pillar2submissionapi/services/SubmitBTNService.scala
@@ -22,8 +22,8 @@ import play.api.libs.json.{JsError, JsSuccess}
 import uk.gov.hmrc.http.{HeaderCarrier, HttpResponse}
 import uk.gov.hmrc.pillar2submissionapi.connectors.SubmitBTNConnector
 import uk.gov.hmrc.pillar2submissionapi.controllers.error.{BTNValidationError, UnexpectedResponse, UnparsableResponse}
-import uk.gov.hmrc.pillar2submissionapi.models.btnsubmissions.BTNSubmission
-import uk.gov.hmrc.pillar2submissionapi.models.btnsubmissions.responses.{SubmitBTNErrorResponse, SubmitBTNSuccessResponse}
+import uk.gov.hmrc.pillar2submissionapi.models.uktrsubmissions.BTNSubmission
+import uk.gov.hmrc.pillar2submissionapi.models.uktrsubmissions.responses.{SubmitBTNErrorResponse, SubmitBTNSuccessResponse}
 
 import scala.concurrent.{ExecutionContext, Future}
