+++ conf/swagger-custom-mappings.yml
@@ -1,0 +1,5 @@
+---
+- type: java\.time\.ZonedDateTime
+  specAsParameter:
+    - type: string
+      format: date-time
