+++ test/uk/gov/hmrc/pillar2submissionapi/controllers/BTNSubmissionControllerSpec.scala
@@ -27,8 +27,8 @@ import uk.gov.hmrc.http.HeaderCarrier
 import uk.gov.hmrc.pillar2submissionapi.base.ControllerBaseSpec
 import uk.gov.hmrc.pillar2submissionapi.controllers.BTNSubmissionControllerSpec._
 import uk.gov.hmrc.pillar2submissionapi.controllers.error.{EmptyRequestBody, InvalidJson}
-import uk.gov.hmrc.pillar2submissionapi.models.btnsubmissions.BTNSubmission
-import uk.gov.hmrc.pillar2submissionapi.models.btnsubmissions.responses.SubmitBTNSuccessResponse
+import uk.gov.hmrc.pillar2submissionapi.models.uktrsubmissions.BTNSubmission
+import uk.gov.hmrc.pillar2submissionapi.models.uktrsubmissions.responses.SubmitBTNSuccessResponse
 
 import scala.concurrent.Future
