+++ test/uk/gov/hmrc/pillar2submissionapi/services/SubmitBTNServiceSpec.scala
@@ -24,8 +24,8 @@ import play.api.test.Helpers.{await, defaultAwaitTimeout}
 import uk.gov.hmrc.http.{HeaderCarrier, HttpResponse}
 import uk.gov.hmrc.pillar2submissionapi.base.UnitTestBaseSpec
 import uk.gov.hmrc.pillar2submissionapi.controllers.error._
-import uk.gov.hmrc.pillar2submissionapi.models.btnsubmissions._
-import uk.gov.hmrc.pillar2submissionapi.models.btnsubmissions.responses.{SubmitBTNErrorResponse, SubmitBTNSuccessResponse}
+import uk.gov.hmrc.pillar2submissionapi.models.uktrsubmissions.BTNSubmission
+import uk.gov.hmrc.pillar2submissionapi.models.uktrsubmissions.responses.{SubmitBTNErrorResponse, SubmitBTNSuccessResponse}
 import uk.gov.hmrc.pillar2submissionapi.services.SubmitBTNServiceSpec._
 
 import java.time.LocalDate
